adding QuizPage component and installed react-router-dom deps
