--- frontend/components/ModalContent.jsx
@@ -1,9 +1,0 @@
-export default function ModalContent({ onClose }) {
-    return (
-      <div className="modal">
-        <div>I'm a modal dialog</div>
-        <button onClick={onClose}>Close</button>
-      </div>
-    );
-  }
-
