+++ frontend/package.json
@@ -18,6 +18,7 @@
     "react": "^19.0.0",
     "react-bootstrap": "^2.10.9",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.3.0",
     "three": "^0.174.0"
   },
   "devDependencies": {
